chore(atomic): add SlotsForNoShadowDOM mixin (#5664)
This PR introduces a new mixin, `SlotsForNoShadowDOMMixin`, to enable
Lit light DOM components to support slot-based content distribution.
This includes methods for adopting child nodes into slots, checking if
slots are empty, and yielding slot content with optional default
fallbacks. in `LitElement` components.

This is necessary to prevent breaking changes in some components, as
this feature exists in Stencil and we are currently using it.

It also renames the file name for the injectSlotsForNoShadowDOM for
better clarity.

https://coveord.atlassian.net/browse/KIT-4498