fix(atomic): remove useless side-effects imports (#5692)
This pull request introduces a modification to the
`externalizeDependenciesPlugin` function in
`packages/atomic/rollup.config.js` to handle side-effect imports from
Headless. The change includes a temporary solution with a TODO comment
for future improvement.

### Key change:
* **Handling side-effect imports**:
- Added a `generateBundle` method to clean up side-effect imports from
Headless by removing specific import statements and extra empty lines in
the generated code. A TODO comment indicates that this is a temporary
solution and suggests exploring better approaches in the future.
  
-----
  
https://coveord.atlassian.net/browse/KIT-4575

---------

Co-authored-by: developer-experience-bot[bot] <91079284+developer-experience-bot[bot]@users.noreply.github.com>