feat(cli): port over create-atomic-component from cli  (#5656)
Continuing the port of atomic-template from CLI.

See https://github.com/coveo/ui-kit/pull/5623 for a full explanation !

This is about porting over the templates that allows to create and
distribute custom atomic components

## `@coveo/create-atomic-rollup-plugin`
Created a `@coveo/create-atomic-rollup-plugin` package to reuse the same
`coveoCdnResolve()` function throughout `@coveo/create-atomic`,
`@coveo/create-atomic-component` and
`@coveo/create-atomic-component-project`.

## `@coveo/create-atomic-component`

`@coveo/create-atomic-component` is a full copy paste from the CLI
repository, with only those changes (so, you can skip over the rest of
the files if you wish, they are not new or modified)

* `package.json` -> Scripts for UI-KIT mono repo (`publish:npm`,
`publish:bump`, `promote:npm:latest`)
* `package.json` -> Repository description
* `project.json` -> nx tasks for `release:phase:1` and `release:phase3`
* `template/sample-component/stencil.config.ts` -> Use `coveoCdnResolve`
rollup plugin
* `template/sample-component/tsconfig.ts` -> Module resolution `bundler`

## `@coveo/create-atomic-component-project`
 
`@coveo/create-atomic-component-project` is a full copy paste from the
CLI repository, with only those changes (so, you can skip over the rest
of the files if you wish, they are not new or modified)


* `package.json` -> Scripts for UI-KIT mono repo (`publish:npm`,
`publish:bump`, `promote:npm:latest`)
* `package.json` -> Repository description
* `template/stencil.config.ts` -> Use `coveoCdnResolve` rollup plugin
* `template/tsconfig.ts` -> Module resolution `bundler`
* `template/index.ts` -> Use `waitForAtomic()` utility function
*  `template/index.ts` -> Remove `organizationEndpoints`
* `template/index.ts` -> Use searchuisamples by default
* `template/index.html` -> Use CDN

## Tests

I * did not * bring back the tests that were present for
`@coveo/create-atomic-component` and
`@coveo/create-atomic-component-project` in the CLI repository.

The reason is that they are done with Verdaccio and jest, which do not
fit with UI-KIT repo.

I can rewrite them as part of this PR if you'd like, but my thinking is
that the PR is already pretty large.

I've done manual test to ensure no regression.

If you prefer I rewrite and bring back the UT as well as part of this
PR, let me know.

[KIT-4508](https://coveord.atlassian.net/browse/KIT-4508)

[KIT-4508]:
https://coveord.atlassian.net/browse/KIT-4508?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Signed-off-by: StepSecurity Bot <bot@stepsecurity.io>
Co-authored-by: renovate-coveo[bot] <115253437+renovate-coveo[bot]@users.noreply.github.com>
Co-authored-by: Louis Bompart <lbompart@coveo.com>
Co-authored-by: vmanek-coveo <vmanek@coveo.com>
Co-authored-by: Alex Prudhomme <78121423+alexprudhomme@users.noreply.github.com>
Co-authored-by: stepsecurity-app[bot] <188008098+stepsecurity-app[bot]@users.noreply.github.com>
Co-authored-by: Felix Perron-Brault <fpbrault@coveo.com>