fix(atomic): add back type declaration (#5746)
I noticed in my recent migration PRs that types _might_ be missing
sometimes oddly.
I think we are missing this?

A git blame brought me to https://github.com/coveo/ui-kit/pull/5017
which had a comment on the matter, but I'm not sure if Stencil do indeed
generate types still.