fix(atomic): always add fileType to the list of fieldsToIncludeInCitations (#5748)
## Issue

When you set a custom `fields-to-include-in-citations` value in the
`atomic-generated-answer` and `atomic-insight-generated-answer`
components, the call to the backend does not include the required fields
for citation anchoring to work (`filetype`).

## Change

In this PR, I edit the logic in order to always add a list of required
fields to `fieldsToIncludeInCitations`, even when customers set a custom
value.

## How to test



https://github.com/user-attachments/assets/46ade44f-0ee4-4a30-a1d7-b65b9b6f593f



This change was tested locally by modifying the `genqa.html` file.

Line 27:
```
await searchInterface.initialize({
        accessToken: '{access-token}',
        organizationId: 'lbergeronsfdevt1z2624x',
        environment:'dev',
        search: {
          pipeline: 'LBergeron RGA Wood',
        },
      }); 
```
Line 119:
```
 <atomic-generated-answer fields-to-include-in-citations="allo, bye"></atomic-generated-answer>
```
[SVCC-5273](https://coveord.atlassian.net/browse/SVCC-5273)

[SVCC-5273]:
https://coveord.atlassian.net/browse/SVCC-5273?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ