feat(quantic): user actions feature integrated in the example insight panel (#5721)
## [SFINT-6174](https://coveord.atlassian.net/browse/SFINT-6174)

In this PR:

- Started using `CoveoV2.PlatformToken.buildInsightTokenPrivileges` in
the InsightTokenProvider class instead pf building the privileges
manually.
- Added the USER_PROFILE privilege with the id of the user who created
the case.
- Introduced a new property to the Quantic Insight Interface component,
the `recordId` this new property will allow the retrieval of the user id
of the user who created the case in order to include the right
privileges in the platform token.
- Added the user actions component to the Example Insight Panel
component.
- Started passing the `Case Id` as a value of the property `Record Id`
to the Quantic Insight Interface in the Example Insight Panel component.



https://github.com/user-attachments/assets/e87c93cf-67a6-4dcc-8581-90c131830a0f

<img width="728" height="818" alt="Screenshot 2025-07-21 at 9 51 38 AM"
src="https://github.com/user-attachments/assets/2cb16d1a-3870-4ae9-bd5d-faa2a887e8a6"
/>


[SFINT-6174]:
https://coveord.atlassian.net/browse/SFINT-6174?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ