fix(headless): populate browser navigator context using relay  (#5680)
**Context**

We recently [improved](https://github.com/coveo-platform/relay/pull/189)
relay to truncate referrer and location urls to with the 1024 limit set
in the schema. However, the locations and referrers sent in API requests
by headless are **not truncated**, which would cause server events with
long urls to be marked as invalid.

**Approach**

The approach covers both CSR and SSR pathways.

The PR adjusts the browser navigator context provider to retrieve the
referrer and location from relay rather than directly by using browser
APIs.

If a custom navigator context provider is provided (e.g., during SSR),
it passes the provider to relay as a custom environment so that relay's
processing is applied. This avoids exposing lower-level utility
functions from the relay package, and reuses a code path used by the
Shopify integration.

**Limitation**

The approach only processes referrers and locations in navigator
context. While locations are only returned by this mechanism, referrers
are sometimes retrieved through a different mechanism (i.e.,
`originLevel3`) and these will not be truncated.

The areas of headless using originLevel3 are in some instances labelled
legacy, but not all. I think it would be better to tackle them in a
separate PR because they raise interesting questions. For example:
- Do we want to deprecate `originLevel3` from state, and make
`navigatorContext` the way to provide referrer?
    - This could make it easier to reason about the codebase.
- It could prevent implementer surprises where specifying one does
nothing because it is not used.
- The `navigatorContext` approach also means that referrer will be
populated in cases where implementers do not set originLevel3
themselves.
- If we do not want to deprecate originLevel3, do we want to truncate
originLevel3 in headless? Or treat it as legacy? Or wait for Data
Platform's position to process urls on the backend?