fix: include enableQuerySyntax in stream-api (#5780)
## Context

This fix follows an investigation into an issue with the Barca Support
site where the AnswerAPI was not generating an answer where the
SearchAPI was.

Using the DevTools to see requests to both endpoints and using Postman
to test requests, it was found that the calls to the AnswerAPI
`/generate` were missing the `enableQuerySyntax` field. Setting this
field allows generating an answer with the particular query that was
raised on the issue.

It turns out that there are tests that covers this section of code, but
the expactions defined in `stream-answer-api-state-mock.ts` did not
include this field.

## Scope

In Stream-Api, if `query` is defined, include `enableQuerySyntax` is the
payload to the generate call.

## Related Jira

[SVCC-5303](https://coveord.atlassian.net/browse/SVCC-5303)

There will be a "fix-it twice" ticket created to include other possible
missing fields.

[SVCC-5303]:
https://coveord.atlassian.net/browse/SVCC-5303?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Signed-off-by: Gilles Leblanc <gleblanc@coveo.com>