fix(headless): bubble api error to the state after renew access token fails (#5589)
https://coveord.atlassian.net/browse/KIT-4455

Errors happening when the renew access token process failed after a 401
or a 419 error were never bubbling up to the state which caused the
state to hang.

This PR fixes that by adding a new action sent from the middleware and
caught in the different slices handling api errors.

I used this regex in vscode to find all the states that had an api error
part : `error:\s*\w*APIError\w*\s*\|\s*null;`