fix(analytics): generativeQuestionAnsweringId (#5787)
https://coveord.atlassian.net/browse/SVCC-5311

# Analytic fix

## The Gist of it

`generativeQuestionAnsweringId` is not the same whether we use the
AnswerAPI or the search api. A selector is taking care to return the
proper value.
This selector must be used everywhere a `generativeQuestionAnsweringId`
is required. It was not the case.

Also, Its always `generativeQuestionAnsweringId` the Backend wants,
never `answerAPIStreamId`

## Side effect.

By introducing a type narrowing in the `SearchAnalyticsProvider` when
introducing the `generativeQuestionAnsweringIdSelector`, We broke the
recommendation pattern, having no search or generatedAnswer section in
its engine state. So need to make sure we are passing the proper
provider to the `configureLegacyAnalytics` function.

## Test coverage

I've tried to cover the change introducted by the selector. + I asked
Claude to find gap in the actual code coveage. Tell be if you think this
is proper.

## How to test manually.

When generating an answer and interacting with the component, the
correct data should be sent to the analytics.
I will try to add a recording soon.