fix(atomic): inject tailwind properties in withTailwindStyles decorator instead of default theme (#5760)
This PR removes the TailwindCSS @property rules from the default theme
(`coveo.css)` and changes the withTailwindStyles decorator so that it
injects them (only once) in the document root instead. This prevents
issues that can happen when modifying/not using the default theme.

This is required due to [@property rules not being supported in shadow
roots](https://github.com/tailwindlabs/tailwindcss/issues/15005).

Additionally it ensures that TailwindCSS @property rules will be
added/removed automatically when updating the TailwindCSS package, as
they are now retrieved dynamically from the `tailwindcss` stylesheet.

https://coveord.atlassian.net/browse/KIT-4634

---------

Co-authored-by: developer-experience-bot[bot] <91079284+developer-experience-bot[bot]@users.noreply.github.com>