fix: Revert "fix(analytics): generativeQuestionAnsweringId (#5787)" (#5811)
This reverts commit d50badefff96841fe85274a43df26f9e05b8e863.

That PR introduced a big circular dependency chain that caused a bug
where the **SearchAnalyticsProvider** class was being called before
being instantiated. This was caught in one of our CI but that action was
not appart of the `is-valid-pr` action. I tried to fix the circular
dependency but this was really complicated and I am not familiar with
the gen AI code. I think we should revert for now.

https://coveord.atlassian.net/browse/KIT-4678