feat(atomic): add logic for pdf citation anchoring with page number (#5788)
[SVCC-5296](https://coveord.atlassian.net/browse/SVCC-5296)

This PR introduces support for citation anchoring to specific pages in
PDF documents when `pageNumber` field is provided in CRGA citations.

⚠️ Note: This feature is currently incomplete since we are still waiting
on receiving the `pageNumber` data

--- 
### What was added

If a citation links to a PDF and a valid `pageNumber` (must be a
positive number) is present, the citation’s URL is updated to include a
`#page={pageNumber}` fragment using the new `generatePdfPageUrl` method.

The `anchorUrl` method was added to divide citation anchoring logic
according to filetype:

- Continues to append a text fragment `:~:text=` for HTML citations
- Appends `#page={pageNumber}` for PDF citations
- Fallback to the regular URL when citation anchoring is disabled, or
when `filetype` and/or `text` are not available.

Feature demoed with hardcoded page number:



https://github.com/user-attachments/assets/2324c2ea-64da-479c-bb76-0337a17a3afd



[SVCC-5296]:
https://coveord.atlassian.net/browse/SVCC-5296?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ