fix(headless): renew expired access token before emitting event through Relay (#5828)
https://coveord.atlassian.net/browse/KIT-4674

## Problem

JWT expiration isn't currently handled for Relay events (fire-and-forget
beacons) in the Headless package.

When dispatching an async thunk action that emits an event through
Relay, an expired JWT currently causes a failure without token renewal /
request retry, since the middleware only handles token renewal
"reactively" after receiving an `UnauthorizedTokenError` (triggered by a
401 or 419 response code, which we can't get from a beacon). While the
JWT would ultimately be renewed in the next async thunk that goes
through the headless platform client, this issue can cause gaps in the
analytics data.

## Proposed solution

- Added a "proactive" JWT check/renewal mechanism in the renew access
token middleware
We check if the access token is expired (or if it will expire within a
60 seconds window), and if so, we try to renew the access token before
resuming the async thunk action.
- Maintain the existing "reactive" token renewal mechanism
If the request still fails with a 401/419, we fallback to the good ol'
token renew / request retry mechanism.

## Testing

The new behavior is covered by unit tests.

In addition, I did manual a manual end-to-end test (using the
headless-commerce-react sample) to ensure that the proposed solution
works.

## Note on performance

The new "proactive" token renewal approach implies that in an
authenticated search scenario, we'll be parsing the JWT and verifying
whether it should be renewed on every dispatched async thunk.

I asked copilot to evaluate the potential performance impact of this
solution, especially as it would add up for scenarios like query suggest
where there can be a high volume of async thunks actions being
dispatched over a short period of time.

Copilot estimates that the performance impact will be negligible:

- JWT format check: ~0.001ms => this is the performance overhead for
every scenario (authenticated search or not)
- JWT parsing (if the access token is a JWT): ~0.03ms => this is the
performance overhead for authenticated search only

Given a typical latency of 50-500ms for network requests, the overhead
would be somewhere between 0.006% - 0.06%, so quite insignificant in the
grand scheme of things.

Regarding query suggest, the impact would still be minimal, even in a
fast typing scenario:

- 5 suggestions/second × 0.03ms = 0.15ms/second overhead
- Over 1 minute of typing = 9ms total overhead
- User types for 10 minutes = 90ms cumulative overhead

We could eventually do some optimization if needed (e.g., storing the
token `exp` in the state to avoid having to parse the JWT multiple
times), but for now I think this would be overkill.