chore(release): seperate workflow name for v2 (#5872)
https://coveord.atlassian.net/browse/KIT-4751

The workflow needs a separate name since now there are two workflow with
the same name but not the same file name in different branches. When
trying to create a release from v2, we get an error since `cd.yml` does
not exist in v2. Having it as a separate name, this means we will see it
in the Actions tab.