docs(atomic): sortCriteria combination for Atomic  (#5855)
When using a sort component, we can specify a list of criteria to apply.
If the first criterion is not enough to sort between two results, we use
the second one, and so forth. However, there are some limitations about
the combination of criteria we can use. This PR covers the changes for
Atomic

---------

Co-authored-by: developer-experience-bot[bot] <91079284+developer-experience-bot[bot]@users.noreply.github.com>