feat: deprecate `AtomicProductText.enableHighlight` and `AtomicCommerceInterface.reflectStateInUrl` (#5911)
We want to adopt HTML Boolean attributes, which must have a default
value of false.
This is a first PR that goes in this direction by deprecating two
boolean attributes that defaulted to true by their semantic
counter-part, with a default to false (so no change in behaviour).

The new attributes are superseding the old ones purposefully: if one use
them/set them, they are "enrolling" in the new behavior.

---------

Co-authored-by: developer-experience-bot[bot] <91079284+developer-experience-bot[bot]@users.noreply.github.com>