fix(headless): analytics with crga (#5857)
[https://coveord.atlassian.net/browse/SVCC-5311](https://coveord.atlassian.net/browse/SVCC-5311)

This PRs:

- Reverts https://github.com/coveo/ui-kit/pull/5811, which had reverted
https://github.com/coveo/ui-kit/pull/5787.
- Fixes the circular dependency that caused the initial revert and had
broken CI.

### Notes

The PR is in two commits:
-
https://github.com/coveo/ui-kit/pull/5857/commits/369f7b6ea853ba42af87fd7a3ab2aa20bb14e608
which puts back the code that was merged in #5811.
-
https://github.com/coveo/ui-kit/pull/5857/commits/e8f9aa0364ff670e39c350758fd92a03311495b3
which breaks the circular dependency.

I suggest reviewing this last commit separately to see what was done in
this PR to address the circular dependency more easily and break the
size of this PR review. You can also review the first commit but note
that it was already reviewed in another PR.

After looking at the various components involved in the circular
dependency, I did not find any easy way to untangle them.
The best path forward I found was to remove the dependency on the
selector to the streamAPI, since the streamAPI (using RTK) is really a
tangled messed of dependencies that can't easily be split and that for
the queries log analytics, which in turns require the selector (at 4 & 5
level deep (2 separate dep paths)).

The selector can now select the value directly from the state, without
resorting to RTK Query. This is an improvement because it's not using
the Query to get a value that's already in state and breaks a circular
dependency loop.

### How to Test

I added breakpoints in the `generated-answer-selectors` in the browser
Dev Tools and testing with both the AnswerAPI and SearchAPI configured
and checked that it was returning the expected value.

I also ran the test suite.

Also modify the `genqa.html` page to:

- add an `answer-configuration-id` like this: `<atomic-generated-answer
answer-configuration-id="fc581be0-6e61-4039-ab26-a3f2f52f308f">"></atomic-generated-answer>`
- add ```analytics: {
          analyticsMode: 'legacy',
        },``` to `searchInterface.initialize`

Test all four combinations to test with AnswerAPI & SearchAPI using the
EP & Legacy Analytics.
The param `generativeQuestionAnsweringId` should be `answerID` when
using the AnswerAPI and `generativeQuestionAnsweringId` with the
SearchAPI.

---------

Co-authored-by: Louis Bompart <lbompart@coveo.com>