chore: update conditions for prerelease jobs in CI workflow (#5972)
This pull request updates the CI workflow conditions for pre-release
jobs to improve build reliability and prevent unnecessary executions.
Now, the pre-release NPM and CDN jobs will only run if the workflow has
not been cancelled, the `is-valid` job succeeded, and the event is a
`merge_group`.

**CI workflow improvements:**

* Updated the condition for the `prerelease-npm-merge` job in
`.github/workflows/ci.yml` to require that the workflow is not
cancelled, the `is-valid` job succeeded, and the event is a
`merge_group`.
* Updated the condition for the `prerelease-cdn-merge` job in
`.github/workflows/ci.yml` with the same checks as above.