fix(atomic): add variable to control number of rows in atomic-commerce-recommendation-list (#5952)
This PR adds a new css variable `--atomic-recs-number-of-rows` that can
be used to control how many rows the
`atomic-commerce-recommendation-list` grid will have.

When the value of `--atomic-recs-number-of-columns` is different than
the productsPerPage value, `--atomic-recs-number-of-rows` can be set to
a value that will ensure the same number of rows is displayed even when
the last page has less products than the others.

Note that while adding --atomic-recs-number-of-rows will help to limit
the layout shifts, it cannot prevent them entirely. This is due to the
fact that the products can have different heights depending on their
content, for example when some sections are hidden unless a field
matches a specific value.

In those cases, a possible workaround is to set a min-height value equal
or larger to the largest atomic-product on the container:

```css
atomic-commerce-recommendation-list::part(result-list-grid-clickable-container) {
	min-height: 200px;
}
```

Alternatively, setting the min-height value on the list itself can also
help.

https://coveord.atlassian.net/browse/KIT-4776