chore(headless-ssr): deprecate `ControllerDefinitionOption` and export missing controller definition types (#5933)
This PR deprecates the `ControllerDefinitionOption` interface and
exports previously missing controller definition types from the index
files.

## Background
The `ControllerDefinitionOption` feature was initially designed to allow
users to specify which solution types (listing, search, standalone,
recommendation) a controller should be built for. For example, a user
could define an InstantProducts controller in the engine definition but
restrict it to only be built for listing contexts, while excluding it
from other solution types.

## Issues with Current Implementation
1. **Out of scope for SSR**: The `ControllerDefinitionOption` feature
extends beyond the intended scope of server-side rendering functionality
2. **Potential confusion**: This option could create confusion for users
about when and how controllers are instantiated as well as the options
to provide
3. **Redundant with existing logic**: Most controllers already have
internal rules that determine whether they should be built based on the
solution type

## Changes Made
- **Deprecated `ControllerDefinitionOption`**: Marked the interface as
deprecated
- **Exported missing types**: Added exports for controller definition
types that were previously missing from the index files

[KIT-4611](https://coveord.atlassian.net/browse/KIT-4611)


[KIT-4611]:
https://coveord.atlassian.net/browse/KIT-4611?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>